refactor: modularize campaign type definitions into separate option and factory modules
