+++ src/types/campaign/Campaign.ts
@@ -9,17 +9,16 @@
  */
 
 import type { IShoppingList } from './acquisition/acquisitionTypes';
-import type { IAutoAwardConfig } from './awards/autoAwardTypes';
+import type { ICampaignOptions } from './CampaignOptions';
 import type { IFactionStanding } from './factionStanding/IFactionStanding';
 import type { SalvageRights, CommandRights } from './Mission';
 import type { ICombatTeam } from './scenario/scenarioTypes';
 
-import { MedicalSystem } from '../../lib/campaign/medical/medicalTypes';
 import { CampaignType } from './CampaignType';
+import { createDefaultCampaignOptions } from './createDefaultCampaignOptions';
 import { MissionStatus, PersonnelStatus } from './enums';
 import { IForce, getAllUnits as getForceUnits } from './Force';
 import { IFinances } from './IFinances';
-import { PersonnelMarketStyle } from './markets/marketTypes';
 import {
   IMission,
   IContract,
@@ -34,356 +33,8 @@ import { IPerson, isActive } from './Person';
 // Re-export Mission types for backwards compatibility
 export type { IMission, IContract, SalvageRights, CommandRights };
 export { isContract, createContract };
-
-// =============================================================================
-// Campaign Options Interface
-// =============================================================================
-
-/**
- * Campaign options for customizing campaign behavior.
- *
- * Contains 40 essential options (of 200+ in MekHQ) organized by category:
- * - Personnel options (~10)
- * - Financial options (~10)
- * - Combat options (~8)
- * - Force options (~6)
- * - General options (~6)
- *
- * @example
- * const options: ICampaignOptions = {
- *   // Personnel
- *   healingRateMultiplier: 1.0,
- *   salaryMultiplier: 1.0,
- *   retirementAge: 65,
- *   // ... more options
- * };
- */
-export type TurnoverFrequency =
-  | 'weekly'
-  | 'monthly'
-  | 'quarterly'
-  | 'annually'
-  | 'never';
-
-export interface ICampaignOptions {
-  // =========================================================================
-  // Personnel Options (~10)
-  // =========================================================================
-
-  /** Multiplier for natural healing rate (1.0 = normal) */
-  readonly healingRateMultiplier: number;
-
-  /** Multiplier for all salaries (1.0 = normal) */
-  readonly salaryMultiplier: number;
-
-  /** Age at which personnel may retire */
-  readonly retirementAge: number;
-
-  /** Days to wait between healing checks */
-  readonly healingWaitingPeriod: number;
-
-  /** Medical system to use (STANDARD, ADVANCED, ALTERNATE) */
-  readonly medicalSystem: MedicalSystem;
-
-  /** Maximum patients per doctor */
-  readonly maxPatientsPerDoctor: number;
-
-  /** Whether doctors use administration skill to increase capacity */
-  readonly doctorsUseAdministration: boolean;
-
-  /** XP awarded per mission */
-  readonly xpPerMission: number;
-
-  /** XP awarded per kill */
-  readonly xpPerKill: number;
-
-  /** Multiplier for skill improvement XP costs (1.0 = normal) */
-  readonly xpCostMultiplier: number;
-
-  /** Whether to track time in service */
-  readonly trackTimeInService: boolean;
-
-  /** Whether to use edge points */
-  readonly useEdge: boolean;
-
-  // =========================================================================
-  // XP Progression Options (~14)
-  // =========================================================================
-
-  /** XP awarded per scenario participation (default 1) */
-  readonly scenarioXP?: number;
-
-  /** XP awarded per kill (default 1) */
-  readonly killXPAward?: number;
-
-  /** Number of kills required to earn killXPAward (default 1) */
-  readonly killsForXP?: number;
-
-  /** XP awarded per task completion (default 1) */
-  readonly taskXP?: number;
-
-  /** Number of tasks required to earn taskXP (default 1) */
-  readonly nTasksXP?: number;
-
-  /** XP awarded per successful vocational training check (default 1) */
-  readonly vocationalXP?: number;
-
-  /** Target number for vocational training 2d6 roll (default 7) */
-  readonly vocationalXPTargetNumber?: number;
-
-  /** Days between vocational training checks (default 30) */
-  readonly vocationalXPCheckFrequency?: number;
-
-  /** XP awarded for administrative duties (default 0) */
-  readonly adminXP?: number;
-
-  /** Days between admin XP awards (default 7) */
-  readonly adminXPPeriod?: number;
-
-  /** XP awarded for mission failure (default 1) */
-  readonly missionFailXP?: number;
-
-  /** XP awarded for mission success (default 3) */
-  readonly missionSuccessXP?: number;
-
-  /** XP awarded for mission outstanding success (default 5) */
-  readonly missionOutstandingXP?: number;
-
-  /** Whether to apply aging effects (attribute decay, trait application) (default true) */
-  readonly useAgingEffects?: boolean;
-
-  // =========================================================================
-  // Financial Options (~10)
-  // =========================================================================
-
-  /** Starting funds in C-bills */
-  readonly startingFunds: number;
-
-  /** Multiplier for maintenance costs (1.0 = normal) */
-  readonly maintenanceCostMultiplier: number;
-
-  /** Multiplier for repair costs (1.0 = normal) */
-  readonly repairCostMultiplier: number;
-
-  /** Multiplier for part acquisition costs (1.0 = normal) */
-  readonly acquisitionCostMultiplier: number;
-
-  /** Whether to pay for unit maintenance */
-  readonly payForMaintenance: boolean;
-
-  /** Whether to pay for repairs */
-  readonly payForRepairs: boolean;
-
-  /** Whether to pay salaries */
-  readonly payForSalaries: boolean;
-
-  /** Whether to pay for ammunition */
-  readonly payForAmmunition: boolean;
-
-  /** Days between maintenance cycles */
-  readonly maintenanceCycleDays: number;
-
-  /** Whether to use loan system */
-  readonly useLoanSystem: boolean;
-
-  /** Whether to use tax system */
-  readonly useTaxes: boolean;
-
-  /** Tax rate as percentage (e.g., 10 = 10%) */
-  readonly taxRate: number;
-
-  /** Overhead percentage of salary (e.g., 5 = 5%) */
-  readonly overheadPercent: number;
-
-  /** Whether to use role-based salary system (monthly via financialProcessor) */
-  readonly useRoleBasedSalaries: boolean;
-
-  /** Whether to pay for secondary role assignments */
-  readonly payForSecondaryRole: boolean;
-
-  /** Maximum loan as percentage of total assets (e.g., 50 = 50%) */
-  readonly maxLoanPercent: number;
-
-  /** Default annual loan interest rate (e.g., 5 = 5%) */
-  readonly defaultLoanRate: number;
-
-  /** Tax payment frequency */
-  readonly taxFrequency: 'monthly' | 'quarterly' | 'annually';
-
-  /** Whether to use food and housing costs */
-  readonly useFoodAndHousing: boolean;
-
-  /** Price multiplier for clan equipment (e.g., 2.0 = 200%) */
-  readonly clanPriceMultiplier: number;
-
-  /** Price multiplier for mixed tech equipment (e.g., 1.5 = 150%) */
-  readonly mixedTechPriceMultiplier: number;
-
-  /** Price multiplier for used equipment (e.g., 0.5 = 50%) */
-  readonly usedEquipmentMultiplier: number;
-
-  /** Price multiplier for damaged equipment (e.g., 0.33 = 33%) */
-  readonly damagedEquipmentMultiplier: number;
-
-  // =========================================================================
-  // Combat Options (~8)
-  // =========================================================================
-
-  /** Whether to use auto-resolve for battles */
-  readonly useAutoResolve: boolean;
-
-  /** Casualty rate multiplier for auto-resolve (1.0 = normal) */
-  readonly autoResolveCasualtyRate: number;
-
-  /** Whether pilots can be captured */
-  readonly allowPilotCapture: boolean;
-
-  /** Whether to use random pilot injuries */
-  readonly useRandomInjuries: boolean;
-
-  /** Chance of pilot death on mech destruction (0.0-1.0) */
-  readonly pilotDeathChance: number;
-
-  /** Whether ejection is automatic on destruction */
-  readonly autoEject: boolean;
-
-  /** Whether to track ammunition usage */
-  readonly trackAmmunition: boolean;
-
-  /** Whether to use quirks system */
-  readonly useQuirks: boolean;
-
-  /** Whether to use AtB dynamic scenario generation (optional, defaults to false) */
-  readonly useAtBScenarios?: boolean;
-
-  /** Difficulty multiplier for OpFor BV calculation (0.5 easy to 2.0 hard, defaults to 1.0) */
-  readonly difficultyMultiplier?: number;
-
-  // =========================================================================
-  // Force Options (~6)
-  // =========================================================================
-
-  /** Maximum units per lance */
-  readonly maxUnitsPerLance: number;
-
-  /** Maximum lances per company */
-  readonly maxLancesPerCompany: number;
-
-  /** Whether to enforce formation rules */
-  readonly enforceFormationRules: boolean;
-
-  /** Whether to allow mixed unit types in formations */
-  readonly allowMixedFormations: boolean;
-
-  /** Whether to require commanders for forces */
-  readonly requireForceCommanders: boolean;
-
-  /** Whether to use combat teams (AtB) */
-  readonly useCombatTeams: boolean;
-
-  // =========================================================================
-  // General Options (~6)
-  // =========================================================================
-
-  /** Date format for display (e.g., 'yyyy-MM-dd') */
-  readonly dateFormat: string;
-
-  /** Whether to use faction-specific rules */
-  readonly useFactionRules: boolean;
-
-  /** Tech level limit (0=Intro, 1=Standard, 2=Advanced, 3=Experimental) */
-  readonly techLevel: number;
-
-  // =========================================================================
-  // Acquisition Options (~4)
-  // =========================================================================
-
-  /** Whether to use the acquisition/procurement system */
-  readonly useAcquisitionSystem?: boolean;
-
-  /** Whether to apply planetary availability modifiers */
-  readonly usePlanetaryModifiers?: boolean;
-
-  /** Time unit for acquisition transit (day, week, or month) */
-  readonly acquisitionTransitUnit?: 'day' | 'week' | 'month';
-
-  /** Whether to apply penalty for Clan parts in IS factions */
-  readonly clanPartsPenalty?: boolean;
-
-  /** Whether to limit equipment by year */
-  readonly limitByYear: boolean;
-
-  /** Whether to allow Clan equipment for IS factions */
-  readonly allowClanEquipment: boolean;
-
-  /** Whether to use random events */
-  readonly useRandomEvents: boolean;
-
-  /** Whether to use prisoner random events */
-  readonly usePrisonerEvents?: boolean;
-
-  /** Whether to use life random events */
-  readonly useLifeEvents?: boolean;
-
-  /** Whether to use contract random events */
-  readonly useContractEvents?: boolean;
-
-  /** Whether to simulate Gray Monday historical event */
-  readonly simulateGrayMonday?: boolean;
-
-  /** Whether to show day report notifications after advancing */
-  readonly enableDayReportNotifications: boolean;
-
-  // =========================================================================
-  // Turnover Options (~8)
-  // =========================================================================
-
-  readonly useTurnover: boolean;
-  readonly turnoverFixedTargetNumber: number;
-  readonly turnoverCheckFrequency: TurnoverFrequency;
-  readonly turnoverCommanderImmune: boolean;
-  readonly turnoverPayoutMultiplier: number;
-  readonly turnoverUseSkillModifiers: boolean;
-  readonly turnoverUseAgeModifiers: boolean;
-  readonly turnoverUseMissionStatusModifiers: boolean;
-
-  // =========================================================================
-  // Faction Standing Options (~2)
-  // =========================================================================
-
-  /** Whether to track faction standing */
-  readonly trackFactionStanding: boolean;
-
-  /** Multiplier for regard changes (1.0 = normal) */
-  readonly regardChangeMultiplier: number;
-
-  // =========================================================================
-  // Auto-Award Options
-  // =========================================================================
-
-  /** Auto-award system configuration (optional, defaults to all enabled) */
-  readonly autoAwardConfig?: IAutoAwardConfig;
-
-  // =========================================================================
-  // Rank System Options
-  // =========================================================================
-
-  /** Code of the active rank system (e.g., 'MERC', 'SLDF', 'CLAN', 'COMSTAR', 'HOUSE') */
-  readonly rankSystemCode?: string;
-
-  // =========================================================================
-  // Market Options
-  // =========================================================================
-
-  /** Unit market generation method ('none' = disabled, 'atb_monthly' = AtB monthly refresh) */
-  readonly unitMarketMethod?: 'none' | 'atb_monthly';
-
-  /** Personnel market generation style */
-  readonly personnelMarketStyle?: PersonnelMarketStyle;
-
-  /** Contract market generation method */
-  readonly contractMarketMethod?: 'none' | 'atb_monthly' | 'cam_ops';
-}
+export type { ICampaignOptions, TurnoverFrequency } from './CampaignOptions';
+export { createDefaultCampaignOptions } from './createDefaultCampaignOptions';
 
 // =============================================================================
 // Campaign Interface
@@ -780,132 +431,6 @@ export function isCampaign(value: unknown): value is ICampaign {
 // =============================================================================
 // Factory Functions
 // =============================================================================
-
-/**
- * Creates default campaign options with sensible defaults.
- *
- * @returns Default ICampaignOptions
- *
- * @example
- * const options = createDefaultCampaignOptions();
- */
-export function createDefaultCampaignOptions(): ICampaignOptions {
-  return {
-    // Personnel options
-    healingRateMultiplier: 1.0,
-    salaryMultiplier: 1.0,
-    retirementAge: 65,
-    healingWaitingPeriod: 1,
-    medicalSystem: MedicalSystem.STANDARD,
-    maxPatientsPerDoctor: 25,
-    doctorsUseAdministration: false,
-    xpPerMission: 1,
-    xpPerKill: 1,
-    xpCostMultiplier: 1.0,
-    trackTimeInService: true,
-    useEdge: true,
-
-    // XP Progression options
-    scenarioXP: 1,
-    killXPAward: 1,
-    killsForXP: 1,
-    taskXP: 1,
-    nTasksXP: 1,
-    vocationalXP: 1,
-    vocationalXPTargetNumber: 7,
-    vocationalXPCheckFrequency: 30,
-    adminXP: 0,
-    adminXPPeriod: 7,
-    missionFailXP: 1,
-    missionSuccessXP: 3,
-    missionOutstandingXP: 5,
-    useAgingEffects: true,
-
-    // Financial options
-    startingFunds: 0,
-    maintenanceCostMultiplier: 1.0,
-    repairCostMultiplier: 1.0,
-    acquisitionCostMultiplier: 1.0,
-    payForMaintenance: true,
-    payForRepairs: true,
-    payForSalaries: true,
-    payForAmmunition: true,
-    maintenanceCycleDays: 7,
-    useLoanSystem: true,
-    useTaxes: true,
-    taxRate: 10,
-    overheadPercent: 5,
-    useRoleBasedSalaries: false,
-    payForSecondaryRole: true,
-    maxLoanPercent: 50,
-    defaultLoanRate: 5,
-    taxFrequency: 'annually',
-    useFoodAndHousing: true,
-    clanPriceMultiplier: 2.0,
-    mixedTechPriceMultiplier: 1.5,
-    usedEquipmentMultiplier: 0.5,
-    damagedEquipmentMultiplier: 0.33,
-
-    // Combat options
-    useAutoResolve: false,
-    autoResolveCasualtyRate: 1.0,
-    allowPilotCapture: true,
-    useRandomInjuries: true,
-    pilotDeathChance: 0.1,
-    autoEject: true,
-    trackAmmunition: true,
-    useQuirks: false,
-
-    // Force options
-    maxUnitsPerLance: 4,
-    maxLancesPerCompany: 3,
-    enforceFormationRules: false,
-    allowMixedFormations: true,
-    requireForceCommanders: false,
-    useCombatTeams: false,
-
-    // General options
-    dateFormat: 'yyyy-MM-dd',
-    useFactionRules: false,
-    techLevel: 1, // Standard
-    limitByYear: true,
-    allowClanEquipment: false,
-    useRandomEvents: false,
-    usePrisonerEvents: true,
-    useLifeEvents: true,
-    useContractEvents: true,
-    simulateGrayMonday: false,
-    enableDayReportNotifications: true,
-
-    // Turnover options
-    useTurnover: true,
-    turnoverFixedTargetNumber: 3,
-    turnoverCheckFrequency: 'monthly',
-    turnoverCommanderImmune: true,
-    turnoverPayoutMultiplier: 12,
-    turnoverUseSkillModifiers: true,
-    turnoverUseAgeModifiers: true,
-    turnoverUseMissionStatusModifiers: true,
-
-    // Faction standing options
-    trackFactionStanding: true,
-    regardChangeMultiplier: 1.0,
-
-    // Acquisition options
-    useAcquisitionSystem: false,
-    usePlanetaryModifiers: true,
-    acquisitionTransitUnit: 'month',
-    clanPartsPenalty: true,
-
-    // Rank system
-    rankSystemCode: 'MERC',
-
-    // Market options
-    unitMarketMethod: 'none',
-    personnelMarketStyle: PersonnelMarketStyle.DISABLED,
-    contractMarketMethod: 'atb_monthly',
-  };
-}
 
 /**
  * Creates a new mission with default values.
