+++ src/types/campaign/createDefaultCampaignOptions.ts
@@ -1,0 +1,111 @@
+import type { ICampaignOptions } from './CampaignOptions';
+
+import { MedicalSystem } from '../../lib/campaign/medical/medicalTypes';
+import { PersonnelMarketStyle } from './markets/marketTypes';
+
+export function createDefaultCampaignOptions(): ICampaignOptions {
+  return {
+    healingRateMultiplier: 1.0,
+    salaryMultiplier: 1.0,
+    retirementAge: 65,
+    healingWaitingPeriod: 1,
+    medicalSystem: MedicalSystem.STANDARD,
+    maxPatientsPerDoctor: 25,
+    doctorsUseAdministration: false,
+    xpPerMission: 1,
+    xpPerKill: 1,
+    xpCostMultiplier: 1.0,
+    trackTimeInService: true,
+    useEdge: true,
+
+    scenarioXP: 1,
+    killXPAward: 1,
+    killsForXP: 1,
+    taskXP: 1,
+    nTasksXP: 1,
+    vocationalXP: 1,
+    vocationalXPTargetNumber: 7,
+    vocationalXPCheckFrequency: 30,
+    adminXP: 0,
+    adminXPPeriod: 7,
+    missionFailXP: 1,
+    missionSuccessXP: 3,
+    missionOutstandingXP: 5,
+    useAgingEffects: true,
+
+    startingFunds: 0,
+    maintenanceCostMultiplier: 1.0,
+    repairCostMultiplier: 1.0,
+    acquisitionCostMultiplier: 1.0,
+    payForMaintenance: true,
+    payForRepairs: true,
+    payForSalaries: true,
+    payForAmmunition: true,
+    maintenanceCycleDays: 7,
+    useLoanSystem: true,
+    useTaxes: true,
+    taxRate: 10,
+    overheadPercent: 5,
+    useRoleBasedSalaries: false,
+    payForSecondaryRole: true,
+    maxLoanPercent: 50,
+    defaultLoanRate: 5,
+    taxFrequency: 'annually',
+    useFoodAndHousing: true,
+    clanPriceMultiplier: 2.0,
+    mixedTechPriceMultiplier: 1.5,
+    usedEquipmentMultiplier: 0.5,
+    damagedEquipmentMultiplier: 0.33,
+
+    useAutoResolve: false,
+    autoResolveCasualtyRate: 1.0,
+    allowPilotCapture: true,
+    useRandomInjuries: true,
+    pilotDeathChance: 0.1,
+    autoEject: true,
+    trackAmmunition: true,
+    useQuirks: false,
+
+    maxUnitsPerLance: 4,
+    maxLancesPerCompany: 3,
+    enforceFormationRules: false,
+    allowMixedFormations: true,
+    requireForceCommanders: false,
+    useCombatTeams: false,
+
+    dateFormat: 'yyyy-MM-dd',
+    useFactionRules: false,
+    techLevel: 1,
+    limitByYear: true,
+    allowClanEquipment: false,
+    useRandomEvents: false,
+    usePrisonerEvents: true,
+    useLifeEvents: true,
+    useContractEvents: true,
+    simulateGrayMonday: false,
+    enableDayReportNotifications: true,
+
+    useTurnover: true,
+    turnoverFixedTargetNumber: 3,
+    turnoverCheckFrequency: 'monthly',
+    turnoverCommanderImmune: true,
+    turnoverPayoutMultiplier: 12,
+    turnoverUseSkillModifiers: true,
+    turnoverUseAgeModifiers: true,
+    turnoverUseMissionStatusModifiers: true,
+
+    trackFactionStanding: true,
+    regardChangeMultiplier: 1.0,
+
+    useAcquisitionSystem: false,
+    usePlanetaryModifiers: true,
+    acquisitionTransitUnit: 'month',
+    clanPartsPenalty: true,
+
+    rankSystemCode: 'MERC',
+
+    unitMarketMethod: 'none',
+    personnelMarketStyle: PersonnelMarketStyle.DISABLED,
+    contractMarketMethod: 'atb_monthly',
+  };
+}
